Fixed a bug in resolve, and resolve now uses buildUri
